Error: Cannot register two commands with the same id: workbench.action.focusCommentsPanel (#202596)
Fixes #202563
